abrindo os arquivos
